requirements.txt
langchain==0.3.6
langchain-community==0.3.4
langchain-groq==0.2.0
langchain-ollama==0.2.0
PyPDF2==3.0.1
tqdm==4.66.5
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
tiktoken==0.8.0
python-dotenv==1.0.1
colorama==0.4.6
